+++ src/__mocks__/@grafana/data.ts
@@ -1,0 +1,12 @@
+const actualData = jest.requireActual('@grafana/data');
+
+const standardEditorsRegistry = {
+  get: () => ({
+    editor: () => null,
+  }),
+};
+
+module.exports = {
+  ...actualData,
+  standardEditorsRegistry,
+};